change back to float32 for cpu

+++ h2o_flexgen/flexgen/pytorch_backend.py
@@ -216,9 +216,9 @@ class TorchDevice:
             # so we only need one workspace instead of two.
             for i in range(1 if policy.sep_layer else 2):
                 shape = (max_seq_len, b * n_head, head_dim)
-                k_cache = self.allocate(shape, np.float16, pin_memory=False)
-                v_cache = self.allocate(shape, np.float16, pin_memory=False)
-                acc = self.allocate(shape[:-1], np.float16, pin_memory=False)
+                k_cache = self.allocate(shape, np.float32, pin_memory=False)
+                v_cache = self.allocate(shape, np.float32, pin_memory=False)
+                acc = self.allocate(shape[:-1], np.float32, pin_memory=False)
                 self.attention_compute_workspace.append((k_cache, v_cache, acc))
         else:
             self.compressed_device.init_attention_compute_workspace(
